Get autolink working.
