--- packages/autolink/module.js
@@ -1,19 +1,0 @@
-/**
- * @copyright   2016-2019, Miles Johnson
- * @license     https://opensource.org/licenses/MIT
- */
-
-// Our index re-exports TypeScript types, which Babel is unable to detect and omit.
-// Because of this, Webpack and other bundlers attempt to import values that do not exist.
-// To mitigate this issue, we need this module specific index file that manually exports.
-
-import Email from './esm/Email';
-import EmailMatcher from './esm/EmailMatcher';
-import Hashtag from './esm/Hashtag';
-import HashtagMatcher from './esm/HashtagMatcher';
-import IpMatcher from './esm/IpMatcher';
-import Link from './esm/Link';
-import Url from './esm/Url';
-import UrlMatcher from './esm/UrlMatcher';
-
-export { Email, EmailMatcher, Hashtag, HashtagMatcher, IpMatcher, Link, Url, UrlMatcher };
