+++ packages/autolink/src/Email.tsx
@@ -1,15 +1,6 @@
 import React from 'react';
-import { ChildrenNode } from 'interweave';
-import Link, { LinkProps } from './Link';
-
-export interface EmailProps extends Partial<LinkProps> {
-  children: ChildrenNode;
-  email: string;
-  emailParts: {
-    host: string;
-    username: string;
-  };
-}
+import Link from './Link';
+import { EmailProps } from './types';
 
 export default function Email({ children, email, emailParts, ...props }: EmailProps) {
   return (
