+++ packages/autolink/src/Url.tsx
@@ -1,20 +1,6 @@
 import React from 'react';
-import { ChildrenNode } from 'interweave';
-import Link, { LinkProps } from './Link';
-
-export interface UrlProps extends Partial<LinkProps> {
-  children: ChildrenNode;
-  url: string;
-  urlParts: {
-    auth: string;
-    fragment: string;
-    host: string;
-    path: string;
-    port: string | number;
-    query: string;
-    scheme: string;
-  };
-}
+import Link from './Link';
+import { UrlProps } from './types';
 
 export default function Url({ children, url, urlParts, ...props }: UrlProps) {
   let href = url;
