+++ packages/autolink/src/index.ts
@@ -3,27 +3,15 @@
  * @license     https://opensource.org/licenses/MIT
  */
 
-import Email, { EmailProps } from './Email';
+import Email from './Email';
 import EmailMatcher from './EmailMatcher';
-import Hashtag, { HashtagProps } from './Hashtag';
+import Hashtag from './Hashtag';
 import HashtagMatcher from './HashtagMatcher';
 import IpMatcher from './IpMatcher';
-import Link, { LinkProps } from './Link';
-import Url, { UrlProps } from './Url';
-import UrlMatcher, { UrlMatcherOptions } from './UrlMatcher';
+import Link from './Link';
+import Url from './Url';
+import UrlMatcher from './UrlMatcher';
 
-export {
-  Email,
-  EmailProps,
-  EmailMatcher,
-  Hashtag,
-  HashtagProps,
-  HashtagMatcher,
-  IpMatcher,
-  Link,
-  LinkProps,
-  Url,
-  UrlProps,
-  UrlMatcher,
-  UrlMatcherOptions,
-};
+export { Email, EmailMatcher, Hashtag, HashtagMatcher, IpMatcher, Link, Url, UrlMatcher };
+
+export * from './types';
