+++ packages/core/package.json
@@ -26,7 +26,7 @@
     "html"
   ],
   "main": "./lib/index.js",
-  "module": "./module.js",
+  "module": "./esm/index.js",
   "types": "./lib/index.d.ts",
   "sideEffects": false,
   "publishConfig": {
