+++ packages/emoji-picker/package.json
@@ -10,7 +10,7 @@
     "picker"
   ],
   "main": "./lib/index.js",
-  "module": "./module.js",
+  "module": "./esm/index.js",
   "types": "./lib/index.d.ts",
   "sideEffects": false,
   "publishConfig": {
