+++ packages/emoji/package.json
@@ -9,7 +9,7 @@
     "emojibase"
   ],
   "main": "./lib/index.js",
-  "module": "./module.js",
+  "module": "./esm/index.js",
   "types": "./lib/index.d.ts",
   "sideEffects": false,
   "publishConfig": {
